requirements.txt
fastapi==0.103.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
uvicorn==0.23.2
httpx==0.24.1
scikit-learn==1.3.2
numpy==1.26.0
pydantic==2.4.2
boto3>=1.28.57
botocore>=1.31.57
typing-extensions==4.8.0